backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
